requirements.txt
six
wheel
tzlocal==2.0.0
PyQt5==5.14.1
pyqtgraph
qdarkstyle
numpy==1.19.3
pandas
matplotlib
seaborn
PyCryptodome==3.9.9
rqdatac==2.9.38
ibapi
deap
pyzmq
wmi
QScintilla==2.11.4
plotly
exchange-calendars
psycopg2-binary

vnpy_rest==1.0.3
vnpy_websocket==1.0.2
vnpy_ctp==6.5.1.7
vnpy_ctptest==6.5.1.3
vnpy_xtp==2.2.27.7
vnpy_tts==6.5.1.1
vnpy_rohon==6.5.1.1
vnpy_mini==1.5.6.3
vnpy_sopt==3.5.5.0
vnpy_uft==3.7.2.4.2
vnpy_esunny==1.0.2.2.3
vnpy_comstar==1.0.1
vnpy_ib==9.81.1.0
vnpy_tap==9.0.0
vnpy_da==1.12.0
vnpy_femas==1.0.0
vnpy_ost==2.0.1.1
vnpy_tora==2021.12.24
vnpy_hft==1.0.0
vnpy_sec==1.0.1
vnpy_hts==1.0.0

vnpy_ctastrategy==1.0.6
vnpy_ctabacktester==1.0.4
vnpy_riskmanager==1.0.2
vnpy_datamanager==1.0.4
vnpy_webtrader==1.0.3
vnpy_spreadtrading==1.1.3
vnpy_datarecorder==1.0.2
vnpy_chartwizard==1.0.1
vnpy_optionmaster==1.0.1
vnpy_algotrading==1.0.1
vnpy_scripttrader==1.0.0
vnpy_portfoliomanager==1.0.0
vnpy_portfoliostrategy==1.0.0
vnpy_paperaccount==1.0.0
vnpy_excelrtd==1.0.0
vnpy_rpcservice==1.0.0

vnpy_sqlite==1.0.0
vnpy_mysql==1.0.0
vnpy_postgresql==1.0.0
vnpy_mongodb==1.0.1
vnpy_influxdb==1.0.1
vnpy_dolphindb==1.0.6
vnpy_leveldb==1.0.1

vnpy_rqdata==2.9.38.1
vnpy_tushare==1.2.64.2
vnpy_tqsdk==2.8.6
vnpy_udata==2021.12.24
vnpy_tinysoft==2021.6.17.2
vnpy_ifind==1.0.0
vnpy_wind==1.0.1
